refactor(a380x/fcdc): move FCDC Publisher to systems/shared
