+++ fbw-a380x/src/systems/instruments/src/EWD/elements/WdAbstractChecklistComponent.tsx
@@ -10,7 +10,7 @@ import {
   VNode,
 } from '@microsoft/msfs-sdk';
 import { EwdSimvars } from '../shared/EwdSimvarPublisher';
-import { FcdcSimvars } from '../../MsfsAvionicsCommon/providers/FcdcPublisher';
+import { FcdcSimvars } from '@shared/publishers/FcdcPublisher';
 import { FwsEvents } from '../../MsfsAvionicsCommon/providers/FwsPublisher';
 import { ChecklistLineStyle, WD_NUM_LINES, WdLineData, WdSpecialLine } from '../../MsfsAvionicsCommon/EcamMessages';
 import { DestroyableComponent } from '../..//MsfsAvionicsCommon/DestroyableComponent';
